fixes package mixup
