+++ app/src/main/java/ikarabulut/http/handlers/MethodNotAllowedHandler.java
@@ -1,8 +1,8 @@
-package ikarabulut.handlers;
+package ikarabulut.http.handlers;
 
-import ikarabulut.response.MethodNotAllowedResponse;
-import ikarabulut.response.Response;
-import ikarabulut.response.StatusCode;
+import ikarabulut.http.response.MethodNotAllowedResponse;
+import ikarabulut.http.response.Response;
+import ikarabulut.http.response.StatusCode;
 
 import java.util.HashMap;
 import java.util.List;
