+++ app/src/main/java/ikarabulut/http/handlers/RequestHandler.java
@@ -1,0 +1,7 @@
+package ikarabulut.http.handlers;
+
+import ikarabulut.http.response.Response;
+
+public interface RequestHandler {
+    Response processResponse();
+}
