+++ app/src/main/java/ikarabulut/http/io/ClientWriter.java
@@ -1,4 +1,4 @@
-package ikarabulut.io;
+package ikarabulut.http.io;
 
 import java.io.*;
 import java.net.Socket;
