+++ app/src/test/java/ikarabulut/http/handlers/MethodNotAllowedHandlerTest.java
@@ -1,6 +1,5 @@
 package ikarabulut.http.handlers;
 
-import ikarabulut.handlers.MethodNotAllowedHandler;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
