+++ app/src/test/java/ikarabulut/http/io/ClientReaderTest.java
@@ -1,6 +1,5 @@
 package ikarabulut.http.io;
 
-import ikarabulut.io.ClientReader;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
