+++ app/src/test/java/ikarabulut/http/io/ClientWriterTest.java
@@ -1,6 +1,5 @@
 package ikarabulut.http.io;
 
-import ikarabulut.io.ClientWriter;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
