+++ app/src/test/java/ikarabulut/http/response/GetResponseTest.java
@@ -1,7 +1,5 @@
 package ikarabulut.http.response;
 
-import ikarabulut.response.GetResponse;
-import ikarabulut.response.StatusCode;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
