+++ app/src/test/java/ikarabulut/http/server/SocketFactoryTest.java
@@ -1,6 +1,5 @@
 package ikarabulut.http.server;
 
-import ikarabulut.server.SocketFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
